Add tca and moved view to module
