+++ App/RootView.swift
@@ -1,0 +1,10 @@
+import Foundation
+import DashFeature
+import SwiftUI
+
+struct RootView: View {
+    var body: some View {
+        DashView()
+    }
+    
+}
