+++ Modules/Package.swift
@@ -7,28 +7,21 @@ let package = Package(
     name: "Modules",
     platforms: [.iOS(.v16)],
     products: [
-        // Products define the executables and libraries a package produces, and make them visible to other packages.
-        .library(
-            name: "Modules",
-            targets: [
-                "DashRepository",
-                "DashRepositoryLive",
-                "NetworkProviders"
-            ]
-        ),
+        .library(name: "DashRepository", targets: ["DashRepository"]),
+        .library(name: "DashRepositoryLive", targets: ["DashRepositoryLive"]),
+        .library(name: "NetworkProviders", targets: ["NetworkProviders"]),
+        .library(name: "DashFeature", targets: ["DashFeature"])
     ],
     dependencies: [
         .package(url: "https://github.com/robbiehanson/CocoaAsyncSocket", branch: "master"),
         .package(url: "https://github.com/pointfreeco/swift-composable-architecture", branch: "main")
-        // Dependencies declare other packages that this package depends on.
-        // .package(url: /* package url */, from: "1.0.0"),
     ],
     targets: [
-        // Targets are the basic building blocks of a package. A target can define a module or a test suite.
-        // Targets can depend on other targets in this package, and on products in packages this package depends on.
         .target(
             name: "DashRepository",
-            dependencies: []
+            dependencies: [
+                .product(name: "Dependencies", package: "swift-composable-architecture")
+            ]
         ),
 
         .target(
@@ -48,7 +41,10 @@ let package = Package(
         
         .target(
             name: "DashFeature",
-            dependencies: []
+            dependencies: [
+                .product(name: "ComposableArchitecture", package: "swift-composable-architecture"),
+                .product(name: "CocoaAsyncSocket", package: "CocoaAsyncSocket")
+            ]
         ),
 
         .testTarget(
