--- Modules/Sources/DashFeature/DraggyView.swift
@@ -1,8 +1,0 @@
-//
-//  File.swift
-//  
-//
-//  Created by Pedro Vargas on 06/01/23.
-//
-
-import Foundation
